Added symbols to toolbuttons

+++ src/main/java/net/osek/survcad/menu/toolbars/DrawToolbar.java
@@ -16,7 +16,7 @@ public class DrawToolbar extends Toolbar {
     }
 
     private void createToolblocks() {
-        long[][] cellOccupancy = new long[][] {
+        int[][] cellOccupancy = new int[][] {
                 {1,1,1,7,7,10,14,16,17},
                 {1,1,1,7,7,8,18,18,19},
                 {1,1,1,9,12,13,18,18,20}};
